Even more refactoring

+++ util/serializer.py
@@ -1,8 +1,10 @@
-import jsonpickle
-from json import JSONEncoder
+import os
 from enum import Enum
-import json
+from pathlib import Path
+import jsonpickle
 import urllib.request
+from app.service import server_default_settings
+from util.settings import default_settings, server_url
 
 path = "data/"
 
@@ -24,8 +26,6 @@ class Serializer:
         self.data_type = data_type
         self.file_name = str(data_type)
 
-
-class SimpleSerializer(Serializer):
     def save(self, data):
         print(f"Saving {self.data_type}...", end=" ")
         for dataType in DataType:
@@ -36,6 +36,10 @@ class SimpleSerializer(Serializer):
 
     def load(self):
         print(f"Loading {self.data_type}...", end=" ")
+        if not isfile(path + self.file_name):
+            print("Done.")
+            # TODO: defaults for every file!!!
+            return default_settings()
         for dataType in DataType:
             if dataType is self.data_type:
                 with open(path + self.file_name, "r") as file:
@@ -46,6 +50,10 @@ class SimpleSerializer(Serializer):
 class ServerSerializer(Serializer):
     def save(self, data):
         print(f"Saving {self.data_type}...", end=" ")
+        full_path = path + self.file_name
+        if not isfile(full_path):
+            print("Done.")
+            return default_settings()
         for dataType in DataType:
             if dataType is self.data_type:
                 with open(path + self.file_name, "w") as file:
@@ -54,26 +62,40 @@ class ServerSerializer(Serializer):
 
     def load(self):
         print(f"Loading {self.data_type}...", end=" ")
-        server_url = "https://gabrielwawerski.github.io/FinanceMate/"
         for dataType in DataType:
             if dataType is self.data_type:
                 with urllib.request.urlopen(f"{server_url}{self.data_type}") as url:
-                    print(f"https://gabrielwawerski.github.io/FinanceMate/{self.data_type}")
                     data = url.read()
                     print("Done.")
                     return jsonpickle.decode(data)
 
 
-class SettingsSerializer(SimpleSerializer):
+class SettingsSerializer(Serializer):
     def __init__(self):
         super().__init__(DataType.SETTINGS)
 
 
-class AccountSerializer(SimpleSerializer):
+class AccountSerializer(Serializer):
     def __init__(self):
         super().__init__(DataType.ACCOUNTS)
 
 
-class TransactionSerializer(SimpleSerializer):
+class TransactionSerializer(Serializer):
     def __init__(self):
         super().__init__(DataType.TRANSACTIONS)
+
+
+def isfile(file_name):
+    return os.path.isfile(file_name)
+
+
+def mkfile(file_name):
+    try:
+        file_path = Path(path + file_name).resolve(strict=True)
+    except FileNotFoundError:
+        if not os.path.isfile(path + file_name):
+            print(f"Loading default settings...")
+            with open(path + file_name, "w+") as f:
+                f.write(server_default_settings())
+    else:
+        print(f"Already exists, omitting: {file_name}")
